Save edits to tags immediately to the server on change...
Summary:
...instead of waiting for the "done" button to be clicked to persist
changes. I've found that it's sooooo easy to forget to click that "Done" button.
If I can't remember to click the "Done" button, then I certainly can't expect
users to do so. It's probably also not obvious that you need to click "Done" for
changes to be saved.

Test Plan:
Click "edit". Remove a tag or add a tag. Refresh page w/o clicking
"Done". Changes should be saved.

Reviewers: spicyj, xymostech

Reviewed By: spicyj

Differential Revision: http://phabricator.benalpert.com/D58