improve search perf vastly and fix issues by searching on in-memory loaded plugins
Summary:
This improves search speed by about 30x on my MBPr, from 1-2 seconds
to consistently about 0.03 seconds for all queries I tried. This also fixes the
issue of query.count() not corresponding with the actual results returned from
query.skip().limit(), due to this RethinkDB bug:
https://github.com/rethinkdb/rethinkdb/issues/1796

(which resulted in the behavior of three results on page 3 of 7 and three
results on page 4 of 7, etc.)

This also makes search keyword-ordering-agnostic, so "python syntax" will match
the same set as "syntax python".

Test Plan:
Tried a bunch of search queries, such as "latex" and "pyt syn", and
sanity-checked results by looking on vim.org. Timed queries using the
@util.time_it decorator. I'm hungry, I'm gonna go eat lunch.

Reviewers: spicyj, xymostech

Reviewed By: spicyj

Differential Revision: http://phabricator.benalpert.com/D106