Load categories from server
Summary:
Add a categories.yaml file that serves as a shared index of all
categories. Read this file on startup and serve it to the frontend via an API
endpoint.

This shared index of categories is useful for pre-caching all tags that belong
to a certain category on app startup. That is still a TODO.

Test Plan:
Refreshed web browser, ensured all categories are still there with
the right icon.

Reviewers: xymostech, spicyj

Reviewed By: spicyj

Differential Revision: http://phabricator.benalpert.com/D159