Add a button to show and edit category on plugin page
Summary:
... to allow people to categorize our un-categorized plugins. Also
shows the category of the plugin and you can click it to see other plugins of
that category (still TODO).

Here's what it looks like:

{F338}

On hover:

{F339}

On click of the edit button, this dropdown fades in:

{F340}

On hover over a select item:

{F341}

Test Plan:
Ensured that the correct category is shown for a bunch of plugins.
Clicked the "edit" button to change category, refreshed, and ensured change is
persisted.

See screenshots above.

Reviewers: xymostech, spicyj

Reviewed By: spicyj

Differential Revision: http://phabricator.benalpert.com/D164