Allow searching by tag using tag:python syntax
Summary:
So you can search for all plugins tagged Python with the search text
"tag:python". The main use of this is for clicking on tags on the sidebar or on
plugin pages, where the link can simply go to a search for that tag.

Test Plan:
Tried "tag:python", "tag:git", etc. Ensured sensible results. Ensured
normal searches still work. Timed response to ensure it's relatively fast and no
regressions.

Reviewers: xymostech, spicyj

Reviewed By: spicyj

Differential Revision: http://phabricator.benalpert.com/D170