Add caching for page 1-10 of empty search and index.html
Summary:
... so we can handle a bit of traffic. Mostly, simultaneous requests to
the homepage.

There's two routes that are cached here: First, the homepage request handler
that responds with the HTML page. Second, the homepage JS makes a GET request to
/api/plugins to populate the search results. This usually takes 150-200ms on my
retina (even for the empty search), which definitely needs to be cached.

Test Plan:
Ran `ab -n 100 '127.0.0.1:5001/api/plugins'` and ab reported ~1000
requests per second (mean).

Ran `ab -n 100 '127.0.0.1:5001/api/plugins?query=&page=11'` and ab reported ~8
requests per second (mean) (this is also definitely less on our server than on
my retina), which is expected because that request is uncached.

Was not able to test just hitting "/" because somehow my localhost seems to always return instantly, cached or not. Will test on vimawesome.com when deployed.

Reviewers: xymostech, spicyj

Reviewed By: spicyj

Differential Revision: http://phabricator.benalpert.com/D192