Fix horizontal scrolling on narrow screens <1200px
Summary:
... with media queries. What happened before was the fixed positioning of the
sidebar would cause it to overlap with the page content with a user attempted to
scroll to the right on a narrow screen. These media queries just removes the
fixed-positioning from the sidebar.

Does not attempt to make layout responsive on small screens, but at least users
can now zoom in w/o the stupid sidebar in the way.

This used to happen before:

{F407}

Now scrolling to the right works:

{F408}

Test Plan: Browsed to all the different pages. Clicked around.

Reviewers: xymostech, spicyj

Reviewed By: spicyj

Differential Revision: http://phabricator.benalpert.com/D200