Improve matching of scraped github plugins with DB
In particular, take into account GitHub's repo ID and the fact that GitHub repo
locations are case-insensitive. Also ensure that GitHub repo scraping is
idempotent -- that is, scraping the same unchanged github repo more than once
should be the same as scraping it once. We used to have a problem of inserting
a new plugin on each scrape, but that is resolved here by saving a foreign key
to the matched plugin in the plugin github repo.