Revert "Use autoprefixer to post-process CSS"
This reverts commit dc6a50e9331724a33b81ecc7e11fdf808d9c5152.

This unfortunately breaks the loading spinner because it doesn't auto-prefix
everything we need. There are also other rules that it doesn't auto-prefix which
we've left in (such as -webkit-font-smoothing), leading to a confusing SCSS
file.

A little too much magic for me. If it were consistent and complete, then maybe.
But in the spirit of making it easier for readers and that code is read more
often than written, I'd elect to keep the vendor prefixes everywhere till we
find a better solution.