mv {clean,review}_submitted_plugins.py
Summary:
Next revision will add a bunch of stuff to this script so that it can actually
be used to interactively review and add plugins instead of just outputting
which submissions need to be reviewied. This just makes the next diff cleaner.

Test Plan: *fingers crossed* (extracted from tested revision, up next)

Reviewers: xymostech, spicyj

Reviewed By: spicyj

Differential Revision: http://phabricator.benalpert.com/D218