Interactively review and add submissions
Summary:
This changes the old script that just displays user submissions that can be
added, and displays a prompt asking whether a submission should be added. If
yes, will insert, else, will record it was rejected (and skip it in future
runs).

This just inserts a row into our table of known plugin github repos, which will
be used when we run the github scraper. A future revision will use the saved
`from_submission` field to auto-populate category and tags.

Test Plan:
- ran `make review_submissions` locally
- responded "y" a bunch of times
- ctrl-c'ed the script
- re-ran. saw that the "yes" submissions were not displayed again and added to
  the count of "known github plugins" and reduced the count of "new plugins".
  Also ran some DB queries to see that new rows were inserted into
  the `plugin_github_repos` table.
- responded "n" a bunch of times
- saw the rejected plugins count go up, new plugins count go down accordingly
- also ran a DB query filering on `rejected: true` to verify

Reviewers: xymostech, spicyj

Reviewed By: spicyj

Differential Revision: http://phabricator.benalpert.com/D217