Bandaid hack: restart RethinkDB (and server) daily
Summary:
... because RethinkDB has been dying
every few days for some reason.  I have tried updating RethinkDB's patch
version, but the problem still seems to be there. There are no errors output to
RethinkDB's log file when this happens.  This may or may not have to do with
Rethinkdb gobbling up too much memory.  For now, we'll just restart the server
and rethinkdb daily -- these restarts are graceful and quick and so shouldn't
be too disruptive.

In the future I should try experimenting with changing Rethinkdb
cache sizes (though it looks like Rethinkdb is already using the minimum
possible memory), or just ask on #rethinkdb.

Test Plan:
- ran script on server
- it worked, things restarted correctly
- hit vimawesome.com to ensure it still works
- saw memory usage drop after restart according to `htop`

Reviewers: spicyj, xymostech

Differential Revision: http://phabricator.benalpert.com/D221