Upgrade rethinkdb to 1.16.1
I was sort of forced to do this after having accidentally upgraded the server's
rethinkdb to 1.14 in September when trying to patch that bash vulnerability.
This caused the data format on the server to not work with 1.13. This came up
as a problem when trying to deploy -- `make init_db` failed because rethinkdb
>= 1.14 removed the `upsert=True` parameter. So I upgraded both the server and
my local rethinkdb version to the latest, 1.16.1, and then fixed all usages of
`upsert=True` to the new `conflict="replace"` argument.

Test Plan:
- ran `make`
- ran `make init_db`
- hit `localhost:5001/`
- ensured things work!