Add simple jest test for component render
react-tools is deprecated but in webpack we're already using jsx-loader, which uses react-tools, so I'm using the same for compatibility. I added this test because it's simple and isolated but as app.jsx is broken up more, it should be easy to add new tests.

For #3.

Test Plan: `npm test`