Revert "Temporarily revert rethinkdb upgrade"
Turns out that despite the fact that we were running rethinkdb version 1.16,
we still had indexes which were created in 1.13.0. There is not a direct
migration path from 1.13.0 indexes straight to 2.2.0. Instead I had to migrate
first to 2.0.5 (Note 2.0.0 SHOULD have worked but there was a bug
https://github.com/rethinkdb/rethinkdb/issues/4775 which we seem to have been
subject to). Once I was on 2.0.5 I was able to rebuild the indexes. From there
I was able to bring us back to 2.2.0 and things worked again.

For the curious, here's the error we were seeing:

    2016-04-15T05:42:01.393253255 0.010778s notice: Running rethinkdb 2.2.0~0trusty (GCC 4.8.2)...
    2016-04-15T05:42:01.407714269 0.025237s notice: Running on Linux 4.4.0-x86_64-linode63 x86_64
    2016-04-15T05:42:01.407772464 0.025290s notice: Loading data from directory /var/lib/rethinkdb/instance1/data
    2016-04-15T05:42:01.419155695 0.036673s notice: Migrating cluster metadata to v2.2
    2016-04-15T05:42:01.875603265 0.493121s error: Found a secondary index from unsupported RethinkDB version 1.13.  You can migrate this secondary index using RethinkDB 2.0.