Being validating plugin submissions
The "Name" and "Author" fields are theoretically required, but up to this point
we have made no effort to actually enforce that requirement. This will give the
user a blocking validation error (in the form of the invalid inputs changing
color) if they try to submit the form with either of these fields empty.

This also sets us up for future validation of the two URL fields.